Use direct message sending for: - `send_debug_to_admin` - `connect_to_forum`
Fix bot handler: processing username.
